% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assn_ibd_rec_pairs.R
\name{assn_ibd_rec_pairs}
\alias{assn_ibd_rec_pairs}
\title{assn_ibd_rec_pairs
given ac and vector of inds in a cluster (cluster1), get inds in each ibd/rec pair}
\usage{
assn_ibd_rec_pairs(ac, cluster1)
}
\description{
assn_ibd_rec_pairs
given ac and vector of inds in a cluster (cluster1), get inds in each ibd/rec pair
}
