% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracKgtX.R
\name{fracKgtX}
\alias{fracKgtX}
\title{fracKgtX.R
for a vector v, returns fraction of values greater than x}
\usage{
fracKgtX(v, x)
}
\description{
fracKgtX.R
for a vector v, returns fraction of values greater than x
}
