% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartitions.R
\name{getPartitions}
\alias{getPartitions}
\title{getPartitions
from allele count (ac), generate list of possible recurrent partitions
nrec_k = number of recurrent allele pairs for partition k
nibd1_k = number of IBD allele pairs in IBD grp1 for partition k
nibd2_k = number of IBD allele pairs in IBD grp1 for partition k
partitions = partition (e.g. 1 for 1:(ac-1), 2 for 2:(ac-2))
nk = number of partitions for this allele count}
\usage{
getPartitions(ac)
}
\description{
getPartitions
from allele count (ac), generate list of possible recurrent partitions
nrec_k = number of recurrent allele pairs for partition k
nibd1_k = number of IBD allele pairs in IBD grp1 for partition k
nibd2_k = number of IBD allele pairs in IBD grp1 for partition k
partitions = partition (e.g. 1 for 1:(ac-1), 2 for 2:(ac-2))
nk = number of partitions for this allele count
}
