% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvals_by_k.R
\name{jvals_by_k}
\alias{jvals_by_k}
\title{function: jvals_by_k
generates matrices of all possible assignments j for each allele pair, for a partition k}
\usage{
jvals_by_k(nk, npairs, partitions)
}
\description{
function: jvals_by_k
generates matrices of all possible assignments j for each allele pair, for a partition k
}
