% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dists.R
\name{load_dists}
\alias{load_dists}
\title{function: load_dists
user provides recombination distances file, if file contains header(T/F), allele count (ac)
distances file format: varID distL_cM distR_cM
this script checks that there are 3 cols, calculate # allele pairs
filter for variants with correct # of allele pairs}
\usage{
load_dists(distsFile, header = T, ac)
}
\description{
function: load_dists
user provides recombination distances file, if file contains header(T/F), allele count (ac)
distances file format: varID distL_cM distR_cM
this script checks that there are 3 cols, calculate # allele pairs
filter for variants with correct # of allele pairs
}
