% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_beta.R
\name{sample_beta}
\alias{sample_beta}
\title{function: sample_beta
sample beta from beta|d,t,alpha}
\usage{
sample_beta(t, alpha, alpha0, beta0)
}
\description{
function: sample_beta
sample beta from beta|d,t,alpha
}
