% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pi.R
\name{sample_pi}
\alias{sample_pi}
\title{function: sample_pi
sampe pi from dirichlet given assignments of k for each variant (vector z), priors (vector piPriors)
nk = number of k values (1=IBD, 1=1:(n-1), 2=2:(n-2), etc.)}
\usage{
sample_pi(z, piPriors, nk)
}
\description{
function: sample_pi
sampe pi from dirichlet given assignments of k for each variant (vector z), priors (vector piPriors)
nk = number of k values (1=IBD, 1=1:(n-1), 2=2:(n-2), etc.)
}
