% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_t.R
\name{sample_t}
\alias{sample_t}
\title{function: sample t (TMRCA)
sample t from gamma distribution, given distance d, alpha, beta
t ~ gamma(alpha,beta)}
\usage{
sample_t(d, alpha, beta)
}
\description{
function: sample t (TMRCA)
sample t from gamma distribution, given distance d, alpha, beta
t ~ gamma(alpha,beta)
}
