% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonIBD_param_values.R
\name{sample_t_nonIBD}
\alias{sample_t_nonIBD}
\title{function: nonIBD_param_values
gibbs sampling of t, beta from recomb distances of known non-IBD allele pairs 
input: dists in format varID, distL, distR
user provided values: n_nonIBDpairs, alpha, alpha0, beta0, niter}
\usage{
sample_t_nonIBD(d, alpha, beta)
}
\description{
function: nonIBD_param_values
gibbs sampling of t, beta from recomb distances of known non-IBD allele pairs 
input: dists in format varID, distL, distR
user provided values: n_nonIBDpairs, alpha, alpha0, beta0, niter
}
