% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_t_z_v.R
\name{sample_t_z_v}
\alias{sample_t_z_v}
\title{function: sample_t_z_v
z = assignments of k for each variant
t = assignments of t for each variant
v = assignments of j for each allele pair
input:
alpha = c(alphaIBD,alphaRec)
beta=c(betaIBD,betaRec)
pi=c(p1,p2,..,pk)
x_i1=distsL
x_i2=distsR
distances are in order so each npairs are for one variant
npairs = numbr of allele pairs for these allele count (=choose(ac,2))}
\usage{
sample_t_z_v(alpha, beta, pi, x_i1, x_i2, npairs)
}
\description{
function: sample_t_z_v
z = assignments of k for each variant
t = assignments of t for each variant
v = assignments of j for each allele pair
input:
alpha = c(alphaIBD,alphaRec)
beta=c(betaIBD,betaRec)
pi=c(p1,p2,..,pk)
x_i1=distsL
x_i2=distsR
distances are in order so each npairs are for one variant
npairs = numbr of allele pairs for these allele count (=choose(ac,2))
}
