% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{function: sampler.r
gibbs sampling of hierarchical model from allele pair recombination distances
user provided values: allele count, alphaRec, betaRec, ..., outfile}
\usage{
sampler(
  dists,
  ac,
  alphaRec,
  betaRec,
  alphaIBD,
  alpha0,
  beta0,
  piPriors,
  niter,
  outfile = "gibbs_sampler_out.rds"
)
}
\description{
function: sampler.r
gibbs sampling of hierarchical model from allele pair recombination distances
user provided values: allele count, alphaRec, betaRec, ..., outfile
}
