% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdRecPairMat.R
\name{ibdRecPairMat}
\alias{ibdRecPairMat}
\title{ibdRecPairMat
for each recurrent partition k for a given allele count (ac), get matrices of all possible IBD/rec pairs}
\usage{
ibdRecPairMat(ac, nibd1_k, nibd2_k, nrec_k)
}
\description{
ibdRecPairMat
for each recurrent partition k for a given allele count (ac), get matrices of all possible IBD/rec pairs
}
