% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dists.R
\name{load_dists}
\alias{load_dists}
\title{load recombination distances file}
\usage{
load_dists(distsFile, header = T, ac)
}
\arguments{
\item{distsFile}{recombination distances file}

\item{header}{logical, does file contains header(T/F)}

\item{ac}{allele count}
}
\value{
tbl of filtered recombination distances
}
\description{
reads in distances file and makes sure it is in the correct format (columns: varID distL_cM distR_cM)
checks that there are 3 cols, calculate # allele pairs
filter for variants with correct # of allele pairs
}
